Use ifName to identify interfaces on unix os
Currently, polling/discovering use "ifIndex" to identify a
port/interface. But on unix distro, the index of a port/interface can
change during a reboot if you remove an interface or if you create them
not in the same order. After a reboot this could create inconsistent
graphs.

I propose to use 'ifName' instead of 'ifIndex' on Linux/BSD distro to
identify the ports/interfaces.